Add NameMixin class
